backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
python-multipart==0.0.6
pymongo==4.15.5